refactor(roguelike_game/game): Creacion de carpeta 'core' dentro de 'game' para posterior refactorizacion y modularizacion
